changed method for retrieving Windows network statistics
Currently, we are running a Powershell method (Get-NetAdapterStatistics) to retrieve the network statistics in case of awsvpc network mode. However, this can cause CPU spikes on smaller instance types. Therefore, we are switching the method to use native Win32 API for retrieving the same.

GetIfTable2Ex is used for the same. Reference: https://learn.microsoft.com/en-us/windows/win32/api/netioapi/nf-netioapi-getiftable2ex

In this refactor, we have abstracted the details of Win32 API in a new package which is used in Windows workflow.