changed method for retrieving Windows network statistics
Currently, we are running a Powershell method (Get-NetAdapterStatistics) to retrieve the network statistics in case of awsvpc network mode. However, this can cause CPU spikes on smaller instance types. Therefore, we are switching the method to use native Win32 API for retrieving the same.

As part of this change, we have added two new APIs to existing NetworkUtils interface applicable for Windows.
- ConvertInterfaceAliasToLUID: Returns the LUID corresponding to the given interface alias.
- GetMIBIfEntryFromLUID: Returns the MIB_IF_ROW for the specified LUID.

Windows API reference:
https://learn.microsoft.com/en-us/windows/win32/api/netioapi/nf-netioapi-convertinterfacealiastoluid
https://learn.microsoft.com/en-us/windows/win32/api/netioapi/nf-netioapi-getifentry2ex