moved firelens capability into common file
We are adding Firelens feature for Windows containers running on Amazon ECS. With this support, Firelens would be supported for both Linux and Windows workloads.

Therefore, we are moving the capabilities related to Firelens into the common file for Linux and Windows. Additionally, we have refactored the tests as well to account for the capability being added to Windows.