windows: execute Set-NetIPAddress cmd for setting lifetime of task IP as infinite
When an instance is launched in private subnet and the task ENI is also launched in private subnet, then the IP address of the task ENI is provided a finite lifetime after which it would be reset to a link-local IP address. To mitigate around this issue, we run the following command from host namespace, which would set the IP address lifetime of task ENI to be infinite.

Reference issue: https://github.com/microsoft/hcsshim/issues/1419