windows: support special characters in the password for FSx
Presently, When a password contains characters that contain escape sequences such as a quotation mark PowerShell fails the command. Reference: https://github.com/aws/amazon-ecs-agent/issues/3270

To mitigate the same and support handling of special characters, we are making the change to use single quoted strings. This means that the password is passed-on to Powershell as it is without any substitutions. Reference:
https://learn.microsoft.com/en-us/powershell/module/microsoft.powershell.core/about/about_quoting_rules?view=powershell-5.1#single-quoted-strings